Add event service on be

+++ backend/src/main/java/com/backend/repositories/UserRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     Optional<User> findByEmail(String email);
 }
